framework/requirements.txt
PyYAML>=6.0
markdown>=3.5
pydantic>=2.0
openai>=1.0
elevenlabs>=0.2
ffmpeg-python>=0.2
google-generativeai>=0.3
python-dotenv>=1.0
google-genai>=0.2

# Browser Automation
selenium>=4.10.0
webdriver-manager>=4.0.0
